fix vulnerability
This commit fixes a severe bug related to the use of Findex. Previously, index
IDs were used to check for user permission upon requesting the application of an
operation. Whenever granted, the operation was applied... without checking this
modification was related to the declared index. Indeed, the index ID was never
used in the derivation of the index addresses.

Due to this absence of verification, any user with write privilege on any index
could ask for the application of an operation on any other index while declaring
the requested modification was performed on its index, which would have resulted
in the application of its requested operation.

The present commit proposes to fix this issue by generating new addresses using
a hash of the addresses in the requested operation and the declared index
ID. That way, the server ensures modifications are tied to the index ID users
declare they are associated to, and thus contain user modifications to the
virtual address space of their index.